+++ com/baidu/tieba/im/message/ResponseDelSystemMessage.java
@@ -3,7 +3,7 @@ package com.baidu.tieba.im.message;
 import com.baidu.adp.framework.message.SocketResponsedMessage;
 import com.squareup.wire.Wire;
 import protobuf.DelGroupMsgs.DelGroupMsgsResIdl;
-/* loaded from: classes.dex */
+/* loaded from: classes2.dex */
 public class ResponseDelSystemMessage extends SocketResponsedMessage {
     private long groupId;
 
@@ -22,7 +22,7 @@ public class ResponseDelSystemMessage extends SocketResponsedMessage {
         setError(delGroupMsgsResIdl.error.errorno.intValue());
         setErrorString(delGroupMsgsResIdl.error.usermsg);
         if (getError() == 0) {
-            this.groupId = delGroupMsgsResIdl.data.groupId.intValue();
+            this.groupId = delGroupMsgsResIdl.data.groupId.longValue();
         }
     }
 }
