+++ com/baidu/tieba/im/message/ResponseSearchGroupMessage.java
@@ -62,7 +62,7 @@ public class ResponseSearchGroupMessage extends SocketResponsedMessage {
             GroupInfo groupInfo = searchGroupResIdl.data.group;
             if (groupInfo != null) {
                 BaseGroupData baseGroupData = new BaseGroupData();
-                baseGroupData.setGroupId(groupInfo.groupId.intValue());
+                baseGroupData.setGroupId(groupInfo.groupId.longValue());
                 baseGroupData.setName(groupInfo.name);
                 baseGroupData.setIntro(groupInfo.intro);
                 baseGroupData.setPortrait(groupInfo.portrait);
