+++ protobuf/CommitInviteMsg/DataRes.java
@@ -4,14 +4,14 @@ import com.squareup.wire.Message;
 import com.squareup.wire.ProtoField;
 /* loaded from: classes.dex */
 public final class DataRes extends Message {
-    @ProtoField(tag = 2, type = Message.Datatype.INT32)
-    public final Integer groupId;
+    @ProtoField(tag = 2, type = Message.Datatype.INT64)
+    public final Long groupId;
     @ProtoField(tag = 1, type = Message.Datatype.INT64)
     public final Long msgId;
     @ProtoField(tag = 3, type = Message.Datatype.INT64)
     public final Long recordId;
     public static final Long DEFAULT_MSGID = 0L;
-    public static final Integer DEFAULT_GROUPID = 0;
+    public static final Long DEFAULT_GROUPID = 0L;
     public static final Long DEFAULT_RECORDID = 0L;
 
     private DataRes(Builder builder, boolean z) {
@@ -42,7 +42,7 @@ public final class DataRes extends Message {
 
     /* loaded from: classes.dex */
     public static final class Builder extends Message.Builder<DataRes> {
-        public Integer groupId;
+        public Long groupId;
         public Long msgId;
         public Long recordId;
