+++ protobuf/CommitReceivedPmsg/DataRes.java
@@ -4,8 +4,8 @@ import com.squareup.wire.Message;
 import com.squareup.wire.ProtoField;
 /* loaded from: classes.dex */
 public final class DataRes extends Message {
-    @ProtoField(tag = 2, type = Message.Datatype.INT32)
-    public final Integer groupId;
+    @ProtoField(tag = 2, type = Message.Datatype.INT64)
+    public final Long groupId;
     @ProtoField(tag = 1, type = Message.Datatype.INT64)
     public final Long readMsgId;
     @ProtoField(tag = 3, type = Message.Datatype.INT64)
@@ -13,7 +13,7 @@ public final class DataRes extends Message {
     @ProtoField(tag = 4, type = Message.Datatype.INT32)
     public final Integer toUserType;
     public static final Long DEFAULT_READMSGID = 0L;
-    public static final Integer DEFAULT_GROUPID = 0;
+    public static final Long DEFAULT_GROUPID = 0L;
     public static final Long DEFAULT_TOUID = 0L;
     public static final Integer DEFAULT_TOUSERTYPE = 0;
 
@@ -51,7 +51,7 @@ public final class DataRes extends Message {
 
     /* loaded from: classes.dex */
     public static final class Builder extends Message.Builder<DataRes> {
-        public Integer groupId;
+        public Long groupId;
         public Long readMsgId;
         public Long toUid;
         public Integer toUserType;
