+++ protobuf/UpgradeMemberGroup/DataReq.java
@@ -2,13 +2,13 @@ package protobuf.UpgradeMemberGroup;
 
 import com.squareup.wire.Message;
 import com.squareup.wire.ProtoField;
-/* loaded from: classes.dex */
+/* loaded from: classes2.dex */
 public final class DataReq extends Message {
     @ProtoField(tag = 2, type = Message.Datatype.INT32)
     public final Integer flag;
-    @ProtoField(tag = 1, type = Message.Datatype.INT32)
-    public final Integer groupId;
-    public static final Integer DEFAULT_GROUPID = 0;
+    @ProtoField(tag = 1, type = Message.Datatype.INT64)
+    public final Long groupId;
+    public static final Long DEFAULT_GROUPID = 0L;
     public static final Integer DEFAULT_FLAG = 0;
 
     private DataReq(Builder builder, boolean z) {
@@ -31,10 +31,10 @@ public final class DataReq extends Message {
         this.flag = builder.flag;
     }
 
-    /* loaded from: classes.dex */
+    /* loaded from: classes2.dex */
     public static final class Builder extends Message.Builder<DataReq> {
         public Integer flag;
-        public Integer groupId;
+        public Long groupId;
 
         public Builder() {
         }
